Scooter do vlastní složky
